requirements.txt
streamlit
google-genai
Pillow
python-dotenv